dom/camera/test/test_camera_bad_initial_config.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bad initial configuration</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id="viewfinder" width="200" height="200" autoplay></video>
<img src="#" alt="This image is going to load" id="testimage"/>
<script class="testbody" type="text/javascript;version=1.7">

var suite = new CameraTestSuite();

suite.test('bad-initial-config', function() {
  function getCamera() {
    var whichCamera = navigator.mozCameras.getListOfCameras()[0];
    var config = {
      mode: 'picture',
      recorderProfile: 'foobar',
    };

    return navigator.mozCameras.getCamera(whichCamera, config);
  }

  function rejectGetCamera(error) {
    ok(error.name === "NS_ERROR_NOT_AVAILABLE",
      "getCamera() failed with: " + error.name);
    return Promise.resolve();
  }

  return getCamera()
    .then(suite.expectedRejectGetCamera, rejectGetCamera);
});

suite.setup()
  .then(suite.run);

</script>
</body>

</html>
